requirements.txt
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
spotipy==2.23.0
urllib3==2.2.1
